Update README.md
OVERVIEW:
The data analysis and machine learning project aimed to analyze a dataset and develop a predictive model. The project involved various steps, including data preprocessing, feature engineering, and evaluation.

KEY POINTS:
The project started with data preprocessing, which involved cleaning the dataset, handling missing values, and transforming variables.
Feature engineering techniques were applied to create new features and improve the predictive power of the model.

The performance of the models was assessed using the evaluation metric r2 score.

The project aims at training the model to predict the Medical Insurance Cost of a person based on their smoking status.
In conclusion, the data analysis and machine learning project successfully analyzed the dataset and developed a predictive model using the linear regression algorithm. 
The project demonstrated the importance of data preprocessing, feature engineering, and model evaluation in achieving accurate predictions. The findings can be used to make informed decisions and improve future predictions in similar domains.